+++ src/update/tests/general.rs
@@ -160,7 +160,7 @@ fn send_raw_key(
     modifiers: iced::keyboard::Modifiers,
     status: iced::event::Status,
 ) -> iced::Task<crate::Message> {
-    app.update(crate::Message::RawKeyEvent(key, modifiers, status))
+    app.update(crate::Message::RawKeyEvent(key, modifiers, status, false))
 }
 
 #[test]